requirements.txt
Flask
Flask_script
Flask_Classy
Flask_SQLAlchemy
pymysql
faker